Release 0.1: Initial version of 'fxMoon'
This release marks the first version of 'fxMoon,' a Discord bot designed to receive user messages and properly display Twitter messages. The bot modifies messages to ensure correct embedding of Twitter links, enhancing the user experience.

Changes:
- Implementation of basic functionality to process user messages.
- Modification of Twitter links in messages for proper display in Discord as it used to be.

Note: This is an initial release, and further enhancements and features are planned for future updates.